requirements.txt
# General dependencies
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.2.1
backoff==2.2.1
boilerpy3==1.0.7
cached-property==1.5.2
canals==0.7.0
cattrs==23.1.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cmake==3.27.6
cryptography==41.0.4
diskcache==5.6.3
docopt==0.6.2
Events==0.5
exceptiongroup==1.1.3
farm-haystack==1.21.0
fastapi==0.103.2
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.9.2
h11==0.14.0
haystack-ai==2.1.1  # Ensure this works with openai >=1.1.0
huggingface-hub==0.17.3
idna==3.4
inflect==7.0.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
lazy-imports==0.3.1
lit==17.0.1
llama_cpp_python==0.2.11
llvmlite==0.41.0
MarkupSafe==2.1.3
monotonic==1.6
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
nltk==3.8.1
num2words==0.5.12
numba==0.58.0
numpy==1.25.2
openai==1.1.0  # Update to the version required by haystack-ai
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
pip==23.2.1
platformdirs==3.11.0
posthog==3.0.2
prompthub-py==4.0.0
pycparser==2.21
pydantic==1.10.13
pynvml==11.5.0
pypdf==3.16.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
quantulum3==0.9.0
rank-bm25==0.2.2
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-cache==0.9.8
rpds-py==0.10.3
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools==68.0.0
six==1.16.0
sniffio==1.3.0
sseclient-py==1.8.0
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken>=0.5.1
torch==2.0.1
tqdm==4.66.1
transformers==4.32.1
typing_extensions==4.8.0
tzdata==2023.3
url-normalize==1.4.3
urllib3==2.0.6
uvicorn==0.23.2
validators==0.22.0
weaviate-client==3.24.1
wheel==0.41.2
yarl==1.9.2


